# Efficiency improvements for VisionSensor