There was a lot of overhead in just the numpy conversion for vision sensors, especially with larger images.  I have made this more efficient.  

Checked that nothing has changed with `assert np.all(img_old_method == img_new_method)`

Comparison of timings:
```
16x16 pix
  RGB Image:
    Original: 0.196ms
    New: 0.026ms
  Depth: 
    Original: 0.068ms
    New: 0.017ms
32x32 pix
  RGB Image:
    Original: 0.482ms
    New: 0.034ms
  Depth:
    Original: 0.201ms
    New: 0.021ms
64x64 pix
  RGB Image:
    Original: 1.025ms
    New: 0.038ms
  Depth:
    Original: 0.366ms
    New: 0.022ms
128x128 pix
  RGB Image:
    Original: 3.233ms
    New: 0.073ms
  Depth:
    Original: 0.939ms
    New: 0.031ms
256x256 pix
  RGB Image:
    Original: 12.530ms
    New: 0.237ms
  Depth:
    Original: 3.666ms
    New: 0.073ms
512x512 pix
  RGB Image:
    Original: 50.817ms
    New: 0.909ms
  Depth:
    Original: 14.615ms
    New: 0.260ms
1024x1024 pix
  RGB Image:
    Original: 213.360ms
    New: 4.615ms
  Depth:
    Original: 62.260ms
    New: 1.074ms
```